Rename `cbor2` to `cbor`
