+++ src/surrealdb/cbor/encoder.py
@@ -1,0 +1,10 @@
+from warnings import warn
+
+from surrealdb.cbor._encoder import CBOREncoder as CBOREncoder
+from surrealdb.cbor._encoder import dump as dump
+from surrealdb.cbor._encoder import dumps as dumps
+from surrealdb.cbor._encoder import shareable_encoder as shareable_encoder
+
+warn(
+    "The cbor.encoder module has been deprecated. Instead import everything directly from cbor."
+)
