+++ src/surrealdb/cbor/types.py
@@ -1,0 +1,17 @@
+from warnings import warn
+
+from surrealdb.cbor._types import CBORDecodeEOF as CBORDecodeEOF
+from surrealdb.cbor._types import CBORDecodeError as CBORDecodeError
+from surrealdb.cbor._types import CBORDecodeValueError as CBORDecodeValueError
+from surrealdb.cbor._types import CBOREncodeError as CBOREncodeError
+from surrealdb.cbor._types import CBOREncodeTypeError as CBOREncodeTypeError
+from surrealdb.cbor._types import CBOREncodeValueError as CBOREncodeValueError
+from surrealdb.cbor._types import CBORError as CBORError
+from surrealdb.cbor._types import CBORSimpleValue as CBORSimpleValue
+from surrealdb.cbor._types import CBORTag as CBORTag
+from surrealdb.cbor._types import FrozenDict as FrozenDict
+from surrealdb.cbor._types import undefined as undefined
+
+warn(
+    "The cbor.types module has been deprecated. Instead import everything directly from cbor."
+)
